requirements.txt
aiohttp==3.1.3
async-timeout==2.0.1
attrs==17.4.0
certifi==2018.4.16
chardet==3.0.4
idna==2.6
idna-ssl==1.0.1
multidict==4.2.0
python-dotenv==0.8.2
requests==2.18.4
telepot==12.6
urllib3==1.22
yarl==1.1.1
